backend/requirements.txt
httpx==0.27.2
yfinance==0.2.43
fastapi-admin
sqlalchemy[asyncio]
jinja2
redis>=5
itsdangerous
sqladmin
sqlalchemy>=2
fastapi
uvicorn[standard]
sqlmodel
psycopg[binary]
pydantic-settings
python-multipart
requests
alembic
EOF